NewsApp V.1/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
newsapi-python
newspaper3k
google-cloud-language
